+++ sphinx/application/data/concepts/repositories/concept-repository-contact/src/main/java/chat/sphinx/concept_repository_contact/ContactRepository.kt
@@ -66,6 +66,10 @@ interface ContactRepository {
         routeHint: LightningRouteHint?
     ): Response<Any, ResponseError>
 
+    suspend fun forceKeyExchange(
+        contactId: ContactId,
+    )
+
     // TODO: add chatId to argument to update alias photo
     suspend fun updateProfilePic(
 //        chatId: ChatId?,